Rename mock metadata to fixture
