+++ npm/test/data/metadata/fixture1.xml
