+++ npm/test/data/metadata/fixture2.js
@@ -1,8 +1,9 @@
+// Fixture for SLO
 module.exports = {
   defaultRedirectUrl: 'http://localhost:3366/sso/oauth/completed',
   redirectUrl: '["http://localhost:3366"]',
-  tenant: 'f0.com',
+  tenant: 'example.com',
   product: 'crm',
-  name: 'testConfig',
-  description: 'Just a test configuration',
+  name: 'SAML Metadata for example.com',
+  description: 'SAML Metadata with SingleLogoutService URLs',
 };
