+++ npm/test/data/metadata/fixture3.js
@@ -1,8 +1,9 @@
+// Fixture without POST/Redirect bindings
 module.exports = {
   defaultRedirectUrl: 'http://localhost:3366/sso/oauth/completed',
   redirectUrl: '["http://localhost:3366"]',
-  tenant: 'example.com',
+  tenant: 'f3.com',
   product: 'crm',
-  name: 'SAML Metadata for example.com',
-  description: 'SAML Metadata with SingleLogoutService URLs',
+  name: 'testConfig',
+  description: 'Just a test configuration',
 };
